Advanced Filter Component added.
